+++ src/Splitio/Services/Parsing/IParser.cs
@@ -1,0 +1,9 @@
+﻿using Splitio.Domain;
+
+namespace Splitio.Services.Parsing.Interfaces
+{
+    public interface IParser<T,P> where T : class where P : class
+    {
+        P Parse(T split);
+    }
+}
