--- src/Splitio/Services/Parsing/Interfaces/ISplitParser.cs
@@ -1,9 +1,0 @@
-﻿using Splitio.Domain;
-
-namespace Splitio.Services.Parsing.Interfaces
-{
-    public interface ISplitParser
-    {
-        ParsedSplit Parse(Split split);
-    }
-}
