+++ src/Splitio/Services/Parsing/Matchers/BaseMatcher.cs
