+++ src/Splitio/Services/Parsing/Matchers/CompareMatcher.cs
