+++ src/Splitio/Services/Parsing/Matchers/EndsWithMatcher.cs
