+++ src/Splitio/Services/Parsing/Matchers/EqualToBooleanMatcher.cs
