+++ src/Splitio/Services/Parsing/Matchers/EqualToSetMatcher.cs
