+++ src/Splitio/Services/Parsing/Matchers/MatchesStringMatcher.cs
